# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Partner API Platform",
        "description": "Partner API + RBAC Admin API. Partner endpoint: POST /api/v1/customers with Bearer token and 1000 req/hour limiter.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application server (paths already include /api/v1 prefix)"
        }
    ],
    "paths": {
...
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List customers",
                "operationId": "bbe0d1a01f9d05b4d5518060da5e7a39",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Ingest customer and payment (partner Bearer token)",
                "operationId": "63608eaeb546cd420e3ce702edc186cf",
                "requestBody": {
                    "$ref": "#/components/requestBodies/PartnerSubmitCustomer"
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Product not available for partner"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "requestBodies": {
            "PartnerSubmitCustomer": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "product_id",
                                "customer_data",
                                "payment"
                            ],
                            "properties": {
                                "partner_id": {
                                    "description": "Partner code. Must match authenticated partner token.",
                                    "type": "string",
                                    "example": "PARTNER_001"
                                },
                                "product_id": {
                                    "description": "Product code enabled for this partner",
                                    "type": "string",
                                    "example": "PROD_123"
                                },
                                "customer_data": {
                                    "description": "Dynamic fields from product configuration with runtime validation.",
                                    "type": "object",
                                    "additionalProperties": true
                                },
                                "payment": {
                                    "required": [
                                        "amount",
                                        "currency",
                                        "paid_at"
                                    ],
                                    "properties": {
                                        "amount": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "USD"
                                        },
                                        "paid_at": {
                                            "description": "ISO-8601 datetime, e.g. 2026-04-17T12:30:00Z",
                                            "type": "string"
                                        },
                                        "transaction_reference": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "metadata": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send header: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}